requirements.txt
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
cryptography==3.1.1
dnspython==2.1.0
greenlet==1.0.0
pyasn1==0.4.8
pycparser==2.20
pymongo==3.11.3
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
pyqt5-tools==5.15.2.3.0.2
python-dotenv==0.15.0
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
rsa==4.6
six==1.15.0
